requirements.txt
asgiref==3.7.2
attrs==23.1.0
autopep8==2.0.2
certifi==2023.5.7
clickhouse-driver==0.2.5
clickhouse-pool==0.5.3
Django==4.2.1
django-clickhouse-backend==1.0.2
django-environ==0.10.0
django-filter==23.2
django-silk==5.0.3
djangorestframework==3.14.0
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.6.1
gprof2dot==2022.7.29
inflection==0.5.1
jsonschema==4.17.3
pycodestyle==2.10.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sentry-sdk==1.25.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.2
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.2
